+++ src/ToolsApi/ToolsApi/Program.cs
@@ -7,7 +7,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace RndStrGen {
+namespace ToolsApi {
     public class Program {
         public static void Main(string[] args) {
             CreateHostBuilder(args).Build().Run();
