+++ src/ToolsApi/ToolsApi/Services/IPService.cs
@@ -1,11 +1,11 @@
 ﻿using Microsoft.AspNetCore.Http;
-using RndStrGen.Interfaces;
+using ToolsApi.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace RndStrGen.Services {
+namespace ToolsApi.Services {
     public class IPService : IIPService {
         public string GetIP(HttpContext context) {
             return context.Connection.RemoteIpAddress.ToString();
